Move `salt.modules.fsutils` to `salt.utils.fsutils`. It's not a salt module.
